protocol/ipv4/tcp - use proper type for length

+++ protocol/ipv4/tcp.cpp
@@ -121,7 +121,7 @@ std::ostream& operator<<( std::ostream& stream, const Header& header ) {
 
 // ** Packet
 
-Packet::Packet( uint8_t& rOctets, size_t len ) {
+Packet::Packet( uint8_t& rOctets, uint16_t len ) {
   m_pHeader_ = new ( &rOctets ) header;
   //m_Content.Init( *m_pHeader_ );
 }
